+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 #### 
 
 ---
+### 0.1.11
+>2020-03-15
+
+#### event addition
+- add event @on-scroll-end , when scroll to bottom trigger it